+++ src/MainPage.js
@@ -1,0 +1,14 @@
+import React from "react";
+
+function MainPage (){
+    return(
+        <>
+        <div>
+            Mainpage
+        </div>
+        </>
+    );
+}
+
+
+export default MainPage;
