+++ src/NotFound.js
@@ -1,0 +1,12 @@
+
+function NotFound(){
+    return (
+        <>
+            <h1>
+                Page Not Found
+            </h1>
+        </>
+    );
+};
+
+export default NotFound;
